requirements.txt
requests[socks]
python-telegram-bot
redis